+++ CHANGELOG.md
@@ -6,14 +6,16 @@
 
 ### Added
 - US for RBAC main screens and navigation accross the app
+- Added filter and order supply batches
 
 ## [0.1.0] - 17/11/2025
 
 ### Changed
  - UploadImageInput modified to not save persistantly the photo, instead to 
    receive the URI value from the view model  
-   
-### Added
+
+## [0.1.0] - 17/11/2025
+
 - US for modify Supply with MVVM and clean architecture principles
 - Delete supply batch
 - Added filter and order products